refactor(backend): colocate punishment domain into feature package
Moves the punishment controllers, entities, repositories, DTOs, expiry
scheduler and Redis-to-Velocity sync classes out of the generic
controller/database/dto/redis/scheduling/utils layer packages into
net.onelitefeather.blackhole.backend.punishment. Renames
PunishmentTemplateHandler to PunishmentTemplateController to match the
naming used by every other domain controller. Pure physical move, no
behavior change.
