+++ backend/src/main/java/net/onelitefeather/blackhole/backend/appeal/AppealEligibilityService.java
@@ -4,13 +4,13 @@ import io.micronaut.context.annotation.Value;
 import jakarta.inject.Singleton;
 import net.onelitefeather.blackhole.backend.database.entities.AppealEntity;
 import net.onelitefeather.blackhole.backend.database.entities.EloProfileEntity;
-import net.onelitefeather.blackhole.backend.database.entities.PunishmentEntity;
 import net.onelitefeather.blackhole.backend.database.repository.AppealRepository;
 import net.onelitefeather.blackhole.backend.database.repository.EloProfileRepository;
 import net.onelitefeather.blackhole.backend.dto.AppealStatus;
 import net.onelitefeather.blackhole.backend.dto.EloTrack;
-import net.onelitefeather.blackhole.backend.dto.PunishType;
 import net.onelitefeather.blackhole.backend.elo.EloService;
+import net.onelitefeather.blackhole.backend.punishment.PunishmentEntity;
+import net.onelitefeather.blackhole.backend.punishment.PunishType;
 import net.onelitefeather.phoca.metadata.Expirable;
 import net.onelitefeather.phoca.metadata.Metadata;
