+++ backend/src/main/java/net/onelitefeather/blackhole/backend/elo/ChatToxicityService.java
@@ -2,11 +2,11 @@ package net.onelitefeather.blackhole.backend.elo;
 
 import io.micronaut.context.annotation.Value;
 import jakarta.inject.Singleton;
-import net.onelitefeather.blackhole.backend.database.entities.PunishmentEvidenceEntity;
-import net.onelitefeather.blackhole.backend.database.repository.PunishmentEvidenceRepository;
 import net.onelitefeather.blackhole.backend.dto.EloReasonCode;
 import net.onelitefeather.blackhole.backend.dto.EloTrack;
-import net.onelitefeather.blackhole.backend.dto.EvidenceType;
+import net.onelitefeather.blackhole.backend.punishment.EvidenceType;
+import net.onelitefeather.blackhole.backend.punishment.PunishmentEvidenceEntity;
+import net.onelitefeather.blackhole.backend.punishment.PunishmentEvidenceRepository;
 import net.onelitefeather.blackhole.backend.utils.SecretHasher;
 
 import java.time.Duration;
