+++ backend/src/main/java/net/onelitefeather/blackhole/backend/elo/EloService.java
@@ -4,15 +4,15 @@ import io.micronaut.context.annotation.Value;
 import jakarta.inject.Singleton;
 import net.onelitefeather.blackhole.backend.database.entities.EloEventEntity;
 import net.onelitefeather.blackhole.backend.database.entities.EloProfileEntity;
-import net.onelitefeather.blackhole.backend.database.entities.PunishmentTemplateEntity;
 import net.onelitefeather.blackhole.backend.database.repository.EloEventRepository;
 import net.onelitefeather.blackhole.backend.database.repository.EloProfileRepository;
-import net.onelitefeather.blackhole.backend.database.repository.PunishmentTemplateRepository;
 import net.onelitefeather.blackhole.backend.dto.EloReasonCode;
 import net.onelitefeather.blackhole.backend.dto.EloTrack;
-import net.onelitefeather.blackhole.backend.dto.PunishType;
 import net.onelitefeather.blackhole.backend.events.DomainEventPublisher;
 import net.onelitefeather.blackhole.backend.punishment.PunishmentApplicationService;
+import net.onelitefeather.blackhole.backend.punishment.PunishmentTemplateEntity;
+import net.onelitefeather.blackhole.backend.punishment.PunishmentTemplateRepository;
+import net.onelitefeather.blackhole.backend.punishment.PunishType;
 import net.onelitefeather.phoca.metadata.Durationable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
