+++ backend/src/main/java/net/onelitefeather/blackhole/backend/imports/VanillaImportService.java
@@ -3,15 +3,15 @@ package net.onelitefeather.blackhole.backend.imports;
 import io.micronaut.core.type.Argument;
 import io.micronaut.json.JsonMapper;
 import jakarta.inject.Singleton;
-import net.onelitefeather.blackhole.backend.database.entities.PunishmentEntity;
-import net.onelitefeather.blackhole.backend.database.entities.PunishmentTemplateEntity;
-import net.onelitefeather.blackhole.backend.database.repository.PunishmentRepository;
-import net.onelitefeather.blackhole.backend.database.repository.PunishmentTemplateRepository;
-import net.onelitefeather.blackhole.backend.dto.PunishType;
 import net.onelitefeather.blackhole.backend.events.DomainEventPublisher;
 import net.onelitefeather.blackhole.backend.profile.CacheInvalidationPublisher;
 import net.onelitefeather.blackhole.backend.profile.PunishmentProfileEntity;
 import net.onelitefeather.blackhole.backend.profile.PunishmentProfileRepository;
+import net.onelitefeather.blackhole.backend.punishment.PunishmentEntity;
+import net.onelitefeather.blackhole.backend.punishment.PunishmentRepository;
+import net.onelitefeather.blackhole.backend.punishment.PunishmentTemplateEntity;
+import net.onelitefeather.blackhole.backend.punishment.PunishmentTemplateRepository;
+import net.onelitefeather.blackhole.backend.punishment.PunishType;
 import net.onelitefeather.blackhole.backend.utils.IdGenerator;
 import net.onelitefeather.blackhole.backend.utils.UUIDHasher;
 import net.onelitefeather.phoca.metadata.Expirable;
