+++ backend/src/main/java/net/onelitefeather/blackhole/backend/profile/PunishProfileDTO.java
@@ -5,7 +5,7 @@ import io.micronaut.core.annotation.Nullable;
 import io.micronaut.core.annotation.ReflectiveAccess;
 import io.micronaut.serde.annotation.Serdeable;
 import jakarta.validation.constraints.Pattern;
-import net.onelitefeather.blackhole.backend.dto.PunishEntryDTO;
+import net.onelitefeather.blackhole.backend.punishment.PunishEntryDTO;
 import net.onelitefeather.phoca.metadata.Metadata;
 
 import java.util.List;
