+++ backend/src/main/java/net/onelitefeather/blackhole/backend/profile/PunishmentProfileEntity.java
@@ -3,7 +3,7 @@ package net.onelitefeather.blackhole.backend.profile;
 import io.micronaut.serde.annotation.Serdeable;
 import jakarta.persistence.*;
 import net.onelitefeather.blackhole.backend.database.converter.MapStringObjectConverter;
-import net.onelitefeather.blackhole.backend.database.entities.PunishmentEntity;
+import net.onelitefeather.blackhole.backend.punishment.PunishmentEntity;
 import org.hibernate.annotations.JdbcTypeCode;
 import org.hibernate.type.SqlTypes;
