+++ backend/src/main/java/net/onelitefeather/blackhole/backend/punishment/EvidenceType.java
@@ -1,4 +1,4 @@
-package net.onelitefeather.blackhole.backend.dto;
+package net.onelitefeather.blackhole.backend.punishment;
 
 /**
  * Where a piece of evidence attached to a punishment originated from.
