+++ backend/src/main/java/net/onelitefeather/blackhole/backend/punishment/PunishmentApplicationService.java
@@ -2,13 +2,8 @@ package net.onelitefeather.blackhole.backend.punishment;
 
 import io.micronaut.context.BeanProvider;
 import jakarta.inject.Singleton;
-import net.onelitefeather.blackhole.backend.database.entities.PunishmentEntity;
-import net.onelitefeather.blackhole.backend.database.entities.PunishmentTemplateEntity;
-import net.onelitefeather.blackhole.backend.database.repository.PunishmentRepository;
-import net.onelitefeather.blackhole.backend.database.repository.PunishmentTemplateRepository;
 import net.onelitefeather.blackhole.backend.dto.EloReasonCode;
 import net.onelitefeather.blackhole.backend.dto.EloTrack;
-import net.onelitefeather.blackhole.backend.dto.PunishType;
 import net.onelitefeather.blackhole.backend.elo.EloService;
 import net.onelitefeather.blackhole.backend.events.DomainEventPublisher;
 import net.onelitefeather.blackhole.backend.profile.CacheInvalidationPublisher;
