+++ backend/src/main/java/net/onelitefeather/blackhole/backend/punishment/PunishmentEvidenceEntity.java
@@ -1,4 +1,4 @@
-package net.onelitefeather.blackhole.backend.database.entities;
+package net.onelitefeather.blackhole.backend.punishment;
 
 import io.micronaut.serde.annotation.Serdeable;
 import jakarta.persistence.Convert;
@@ -10,8 +10,6 @@ import jakarta.persistence.Index;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import net.onelitefeather.blackhole.backend.database.converter.MapStringObjectConverter;
-import net.onelitefeather.blackhole.backend.dto.EvidenceType;
-import net.onelitefeather.blackhole.backend.dto.PunishmentEvidenceDTO;
 import org.hibernate.annotations.JdbcTypeCode;
 import org.hibernate.type.SqlTypes;
